uno/requirements.txt
# Core LangChain packages
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0

# Vector database
faiss-cpu==1.8.0

# Optional: Use faiss-gpu instead if you have GPU support
# faiss-gpu==1.8.0

# Document processing
pypdf==4.3.1
python-dotenv==1.0.1

# Development and utility
python-dotenv
